+++ packages/bff/src/application/index.ts
@@ -2,3 +2,4 @@ export * from "./orchestrator/aggregation.service";
 export * from "./orchestrator/mutation-orchestrator.service";
 export * from "./orchestrator/query-orchestrator.service";
 export * from "./orchestrator/query-pipeline.service";
+export * from "./view/temporary-view-adapter.service";