+++ packages/bff/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./shared.types";
 export * from "./config";
+export * from "./view.types";